flags: refactor and rename
The old name ("set-flags") implied that it could only set flags even
though it was also able to clear them.  The new name ("flags")
suggests that it works with flags.

I also changed the syntax for setting and clearing flags.  The new
syntax allows the command to support arbitrary, single-character flags
while still allowing the "flags" command to accept other command line
arguments in the future.