count: allow query arguments
Now that count and find use Find() under the hood, they can both
accept the same arguments for restricting the query results.

The changes to mailz.sh show a complicated output format using the
results of three different counts with different flags.  I won't stick
with this format, but it highlights how it might be used.