head: support time headers
I tested this with both Date and Received headers which contain
timestamps in different locations.

I chose RFC 3339 time format since it's unambiguous, it contains no
space characters (which can be confusing for Unix utilities), and it
sorts in chronological order.  Tools further down a pipeline can
reformat the timestamp however the user wants it.  If a user doesn't
like these decisions, he can always use "-s" to get the raw string
content and deal with it himself.