head: make -N replace tabs with spaces
The -N type is intended to be a high-level convenience.  Including raw
tab characters in its output goes against that spirit of convenience.
By expanding to 8 spaces, the essence of the tab character remains.

Users who need the tab character can use -s instead of -N and parse
the name according to their own preferences.