example: split generating and rendering message list
When displaying a message list, there are two steps: generate the raw
message list, and render that list into something a user wants to see.

By keeping those two steps separate, I can render subsets of the list
or regenerate the list without rendering it or make a minor
modification to the list without traversing all messages again (such
as changing which messages are selected).