head: -i flag for including message ID
The message ID (the unique portion of a maildir message filename) is a
stable identifier which can be given to mailz commands to refer to a
specific message.  Including it in head's output can be helpful for
scripting.  For example, "find all messages where Subject has 'foo'"
becomes something like:

    mailz head -i -s Subject ... | grep foo | cut -f 1