example: only generate message list if necessary
When producing a message list, only generate it from scratch if its
missing.  This allows users to view the list while retaining their
current selection (previously the selection was lost).

There's a new 'L' binding which forces the list to be generated from
scratch.  That can be helpful after messages are marked as done or
expunged.