body: handle multipart/mixed messages better
A body with type multipart/mixed requires that we display all of the
parts.  Previously, we displayed only the first part.

This new code recursively outputs the body of a message.  For
multipart/alternative, it displays the first viewable part.  For
multipart/mixed, it displays all viewable parts.  It also gives an
error if there are no viewable parts.

This doesn't happen often, but this code should be able to handle a
multipart/alternative message with nested multipart/mixed messages
inside it.