body: display attachments
I'm not entirely sure that this is the right thing to do.  On the one
hand, it's helpful to be able to see when an attachment exists in the
body, even if its content can't be displayed.  On the other hand, this
kind of output is indistinguishable from text written in a text/plain
body.  That makes it difficult to script.

I lean towards including this change (obviously) since users who need
precise control over each part of a message body will have to process
the raw mail file directly.  Those who don't need that control
probably value the attachment summary.