move: new command
The move command is just syntactic sugar around cur, copy and flags.
However, the move operation is very helpful and it's likely that users
will get it wrong if they have to reimplement it themselves
repeatedly.  Our example script uses it once but will use it again
shortly.  Making it a built in operation seems best for everyone.