example: correctly sort by Subject and then Received
The -k option specifies a starting and ending field for sorting.  If
the ending field is omitted, it defaults to the last field in the
line.  I assumed it defaulted to the starting line.  We only want the
sort key to include the specified line, so we have to explicitly make
it the starting and ending field.